+++ src/ast/desugar.rs
@@ -107,6 +107,7 @@ pub(crate) fn desugar_command(
                         Schema {
                             input: variant.types,
                             output: datatype.clone(),
+                            extra_outputs: vec![],
                         },
                         variant.cost,
                         false,
@@ -242,6 +243,7 @@ fn desugar_prove(parser: &mut Parser, span: Span, query: Vec<Fact>) -> Vec<NComm
             Schema {
                 input: vec![],
                 output: fresh_sort.clone(),
+                extra_outputs: vec![],
             },
             None,
             false,
@@ -294,6 +296,7 @@ fn desugar_datatype(span: Span, name: String, variants: Vec<Variant>) -> Vec<NCo
             Schema {
                 input: variant.types,
                 output: name.clone(),
+                extra_outputs: vec![],
             },
             variant.cost,
             variant.unextractable,
@@ -418,6 +421,7 @@ fn desugar_relation(
             Schema {
                 input: inputs,
                 output: fresh_sort,
+                extra_outputs: vec![],
             },
             None,
             false,