+++ src/proofs/proof_encoding.rs
@@ -903,6 +903,9 @@ impl<'a> ProofInstrumentor<'a> {
 
                         (fv.clone(), proof)
                     }
+                    ResolvedCall::Values(_) => {
+                        panic!("tuple-output (`values`) functions are not supported in proofs")
+                    }
                 }
             }
         }
@@ -1139,6 +1142,9 @@ impl<'a> ProofInstrumentor<'a> {
                             ListDisplay(args, " ")
                         ));
                         fv
+                    }
+                    ResolvedCall::Values(_) => {
+                        panic!("tuple-output (`values`) functions are not supported in proofs")
                     }
                 }
             }